# swagger/docs/swagger.yaml
definitions:
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  handlers.WalletOperationRequest:
    properties:
      amount:
        type: number
      currency:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/balance:
    get:
      description: Get current balance of user's wallet
      produces:
      - application/json
      responses:
        "200":
          description: Wallet balance
          schema:
            additionalProperties:
              type: number
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Wallet not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get wallet balance
      tags:
      - wallet
  /v1/create-wallet:
    post:
      consumes:
      - application/json
      description: Create a new wallet for authenticated user
      produces:
      - application/json
      responses:
        "201":
          description: Wallet created successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new wallet
      tags:
      - wallet
  /v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /v1/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email and password
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: User already exists
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /v1/wallet/deposit:
    post:
      consumes:
      - application/json
      description: Deposit money into user's wallet
      parameters:
      - description: Deposit details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.WalletOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Operation successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deposit money
      tags:
      - wallet
swagger: "2.0"